+++ graal-js/src/com.oracle.truffle.js/src/com/oracle/truffle/js/runtime/JSContext.java
@@ -285,6 +285,10 @@ public class JSContext {
         SharedArrayBufferGetByteLength,
         FunctionAsyncIterator,
         IsGraalRuntime,
+        AsyncModuleExecutionFulfilled,
+        AsyncModuleExecutionRejected,
+        TopLevelAwaitResolve,
+        TopLevelAwaitReject,
     }
 
     @CompilationFinal(dimensions = 1) private final JSFunctionData[] builtinFunctionData;