+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The main focus is on user-observable behavior of the engine.
 ## Version 20.2.0
 * Implemented the [Intl.NumberFormat Unified API](https://github.com/tc39/proposal-unified-intl-numberformat) proposal.
 * Implemented the [Logical Assignment Operators](https://github.com/tc39/proposal-logical-assignment) proposal.
+* Implemented the [Top-level Await](https://github.com/tc39/proposal-top-level-await) proposal.
 
 ## Version 20.1.0
 * ECMAScript 2020 mode/features enabled by default.